Move ImageRequestKeys to an external file and reduce access to CoW container
